ServerHandler.java -> ServerEvents.java to match ClientEvents.java
Moved contentsHaveEffect to an ItemEffect static method as it is also used in ServerEvents.
Fixed tossing contents of an item container instead of the item container itself. Required a significant rework
